HandleException

+++ src/main/java/com/monkey/controller/ProductServiceController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.monkey.exception.ProductNotFoundException;
 import com.monkey.model.Product;
 
 @RestController
@@ -41,6 +42,7 @@ public class ProductServiceController {
 	
 	@RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
 	public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product) {
+		if(!productRepo.containsKey(id)) throw new ProductNotFoundException();
 		productRepo.remove(id);
 		product.setId(id);
 		productRepo.put(id, product);
@@ -49,6 +51,7 @@ public class ProductServiceController {
 	
 	@RequestMapping(value = "products/{id}", method = RequestMethod.DELETE)
 	public ResponseEntity<Object> deleteProduct(@PathVariable("id") String id) {
+		if(!productRepo.containsKey(id)) throw new ProductNotFoundException();
 		productRepo.remove(id);
 		return new ResponseEntity<>("Product is deleted successfully", HttpStatus.OK);
 	}
